waterly_server/backend/requirements.txt
fastapi
uvicorn
paho-mqtt
requests